+++ src/audiodevice.rs
@@ -95,6 +95,10 @@ impl ChunkStats {
             .collect()
     }
 
+    pub fn rms_linear(&self) -> Vec<f32> {
+        self.rms.iter().map(|val| *val as f32).collect()
+    }
+
     pub fn peak_db(&self) -> Vec<f32> {
         self.peak
             .iter()
@@ -106,6 +110,10 @@ impl ChunkStats {
                 }
             })
             .collect()
+    }
+
+    pub fn peak_linear(&self) -> Vec<f32> {
+        self.peak.iter().map(|val| *val as f32).collect()
     }
 }