+++ src/cpaldevice.rs
@@ -378,10 +378,8 @@ impl PlaybackDevice for CpalPlaybackDevice {
                                         }
                                     }
                                     chunk_stats = chunk.get_stats();
-                                    playback_status.write().unwrap().signal_rms =
-                                        chunk_stats.rms_db();
-                                    playback_status.write().unwrap().signal_peak =
-                                        chunk_stats.peak_db();
+                                    playback_status.write().unwrap().signal_rms.add_record_squared(chunk_stats.rms_linear());
+                                    playback_status.write().unwrap().signal_peak.add_record(chunk_stats.peak_linear());
                                     tx_dev.send(chunk).unwrap();
                                 }
                                 Ok(AudioMessage::Pause) => {
@@ -672,8 +670,8 @@ impl CaptureDevice for CpalCaptureDevice {
                             }
                             chunk_stats = chunk.get_stats();
                             //trace!("Capture rms {:?}, peak {:?}", chunk_stats.rms_db(), chunk_stats.peak_db());
-                            capture_status.write().unwrap().signal_rms = chunk_stats.rms_db();
-                            capture_status.write().unwrap().signal_peak = chunk_stats.peak_db();
+                            capture_status.write().unwrap().signal_rms.add_record_squared(chunk_stats.rms_linear());
+                            capture_status.write().unwrap().signal_peak.add_record(chunk_stats.peak_linear());
                             value_range = chunk.maxval - chunk.minval;
                             state = silence_counter.update(value_range);
                             if state == ProcessingState::Running {