+++ src/helpers.rs
@@ -68,3 +68,14 @@ pub fn multiply_add_elements(
         *r += val;
     }
 }
+
+// Inplace recalculation of values positive values 0..1 to dB.
+pub fn linear_to_db(values: &mut [f32]) {
+    values.iter_mut().for_each(|val| {
+        if *val == 0.0 {
+            *val = -1000.0;
+        } else {
+            *val = 20.0 * val.log10();
+        }
+    });
+}