+++ src/pulsedevice.rs
@@ -197,10 +197,16 @@ impl PlaybackDevice for PulsePlaybackDevice {
                                             conversion_result.1;
                                     }
                                     chunk_stats = chunk.get_stats();
-                                    playback_status.write().unwrap().signal_rms =
-                                        chunk_stats.rms_db();
-                                    playback_status.write().unwrap().signal_peak =
-                                        chunk_stats.peak_db();
+                                    playback_status
+                                        .write()
+                                        .unwrap()
+                                        .signal_rms
+                                        .add_record_squared(chunk_stats.rms_linear());
+                                    playback_status
+                                        .write()
+                                        .unwrap()
+                                        .signal_peak
+                                        .add_record(chunk_stats.peak_linear());
                                     //trace!(
                                     //    "Playback signal RMS: {:?}, peak: {:?}",
                                     //    chunk_stats.rms_db(),
@@ -422,8 +428,8 @@ impl CaptureDevice for PulseCaptureDevice {
                                     break;
                                 }
                             }
-                            capture_status.write().unwrap().signal_rms = chunk_stats.rms_db();
-                            capture_status.write().unwrap().signal_peak = chunk_stats.peak_db();
+                            capture_status.write().unwrap().signal_rms.add_record_squared(chunk_stats.rms_linear());
+                            capture_status.write().unwrap().signal_peak.add_record(chunk_stats.peak_linear());
                         }
                         capture_status.write().unwrap().state = ProcessingState::Inactive;
                     }